Door update

+++ src/main_pack/Game.java
@@ -94,8 +94,8 @@ public class Game implements Runnable {
 
         initLootTables();
         player = new Player(1, 3, this);
-        camera.setX(-player.getPixelPosition(player.getX()) + Launcher.WIDTH / 2 - (int) (UNIT_SCALE) / 2);
-        camera.setY(-player.getPixelPosition(player.getY()) + Launcher.HEIGHT / 2 - (int) (UNIT_SCALE) / 2);
+        camera.setX(-player.getPixelPosition(player.getX()) + Launcher.WIDTH / 2 - UNIT_SCALE / 2);
+        camera.setY(-player.getPixelPosition(player.getY()) + Launcher.HEIGHT / 2 - UNIT_SCALE / 2);
 
 
         //GameState Classes
